+++ NodeJS/codeExamples/dataArrayOfObject.js
@@ -1,0 +1,35 @@
+const data = [
+  {
+    testingNewValue: {
+      "en-US": "hello",
+    },
+  },
+
+  {
+    testingNewValue: {
+      "hi-IN": "hello ",
+    },
+  },
+
+  {
+    testingNewValue: {
+      us: "dummy",
+    },
+  },
+
+  {
+    testingNewValue: {
+      "xp-op-as": "value for testing",
+    },
+
+    locationField: {
+      "en-US": {
+        lat: 19.28563,
+
+        lon: 72.8691,
+      },
+    },
+  },
+];
+
+module.exports = {data};
