+++ NodeJS/codeExamples/dataObj.js
@@ -1,36 +1,8 @@
- const data = [
-    {
+ const data = {
       testingNewValue: {
         "en-US": "hello",
       },
-    },
-  
-    {
-      testingNewValue: {
-        "hi-IN": "hello ",
-      },
-    },
-  
-    {
-      testingNewValue: {
-        us: "dummy",
-      },
-    },
-  
-    {
-      testingNewValue: {
-        "xp-op-as": "value for testing",
-      },
-  
-      locationField: {
-        "en-US": {
-          lat: 19.28563,
-  
-          lon: 72.8691,
-        },
-      },
-    },
-  ];
+    };
   
 // module => commonjs module
 module.exports = { data };
