+++ NodeJS/codeExamples/objectOperations.js
@@ -1,8 +1,3 @@
 const data = require('./dataObj');
+
 console.log(data.data);
-/*
-const callback = function (obj, index) {
-  console.log(`index ${index}`, Object.keys(obj.testingNewValue)[0]);
-};
-data.data.forEach(callback);
-*/
