+++ NodeJS/codeExamples/package.json
@@ -7,4 +7,4 @@
     "scripts": {
       "test": "echo \"Error: no test specified\" && exit 1"  },  "author": "Chikeluba Anusionwu",
     "license": "ISC"
-  }
+}
